refactor: remove dom.js barrel, import directly from sub-modules
All 50 consumers now import from dom-core.js, dom-buttons.js,
dom-tabs.js, or dom-lists.js instead of the barrel re-export dom.js.
This eliminates the coupling hotspot (34 dependents → 0) and makes
each file's actual dependencies explicit.

Test mocks updated accordingly.
