+++ src/components/board-view.js
@@ -1,7 +1,9 @@
 import {
   onTerminalCreated, onTerminalRemoved, onTerminalExited,
 } from '../utils/terminal-events.js';
-import { _el, buildDomainButtonBar, renderList } from '../utils/dom.js';
+import { _el } from '../utils/dom-core.js';
+import { buildDomainButtonBar } from '../utils/dom-buttons.js';
+import { renderList } from '../utils/dom-lists.js';
 import {
   _safeFit, disposeTerminal, disposeTerminalMap, setupTerminalAddons,
   createPtyBoundTerminal,
