+++ src/components/flow-view.js
@@ -1,7 +1,7 @@
 import { registerComponent, getComponent } from '../utils/component-registry.js';
 import { ComponentBase } from '../utils/component-base.js';
 import { moveFlowInOrder } from '../utils/flow-view-helpers.js';
-import { toggleCollapsible } from '../utils/dom.js';
+import { toggleCollapsible } from '../utils/dom-lists.js';
 import {
   addCategory, renameCategoryInline, deleteCategory,
 } from '../utils/flow-view-categories.js';
