+++ src/components/settings-modal.js
@@ -1,5 +1,7 @@
 import { formatCombo, eventToCombo } from '../utils/shortcut-helpers.js';
-import { _el, createActionButton, buildTabBar } from '../utils/dom.js';
+import { _el } from '../utils/dom-core.js';
+import { createActionButton } from '../utils/dom-buttons.js';
+import { buildTabBar } from '../utils/dom-tabs.js';
 import { createModalOverlay } from '../utils/dom-dialogs.js';
 import { MODAL_CLOSE_TRANSITION_MS, MODIFIER_KEYS, NAV_SECTIONS } from '../utils/settings-helpers.js';
 import { getComponent } from '../utils/component-registry.js';
