+++ src/utils/file-editor-renderer.js
@@ -3,7 +3,7 @@
  * Extracted from file-viewer.js to reduce component size.
  */
 
-import { _el } from './dom.js';
+import { _el } from './dom-core.js';
 import { renderStatusBar } from './view-header.js';
 import { getCursorPosition, insertTab, SAVE_FLASH_MS, TAB_SPACES } from './editor-helpers.js';
