+++ src/utils/file-viewer-tabs.js
@@ -3,7 +3,8 @@
  * Extracted from file-viewer.js to reduce component size.
  */
 
-import { _el, renderList } from './dom.js';
+import { _el } from './dom-core.js';
+import { renderList } from './dom-lists.js';
 import { attachContextMenu } from './context-menu.js';
 import { createTabElement } from './tab-renderer.js';
