+++ src/utils/flow-card-setup.js
@@ -8,7 +8,8 @@
  * @typedef {{ id: string, runs?: Array<FlowRun>, enabled?: boolean }} FlowDescriptor
  */
 
-import { _el, toggleCollapsible } from './dom.js';
+import { _el } from './dom-core.js';
+import { toggleCollapsible } from './dom-lists.js';
 import { getLastRun } from './flow-view-helpers.js';
 import { cleanupAllDragState } from './flow-drag-cleanup.js';
 import { createCardHeader } from './flow-card-renderer.js';
