+++ src/utils/flow-category-renderer.js
@@ -3,8 +3,9 @@
  * Handles category headers, collapse state, and drag-drop zone setup.
  * Extracted from flow-view.js to reduce component size.
  */
-import { _el, buildDomainButtonBar } from './dom.js';
-import { buildChevronRow } from './dom.js';
+import { _el } from './dom-core.js';
+import { buildDomainButtonBar } from './dom-buttons.js';
+import { buildChevronRow } from './dom-lists.js';
 import { setupDropZone } from './drop-zone-helpers.js';
 import { CATEGORY_ACTIONS } from './flow-view-helpers.js';
 import { computeInsertionIndex } from './drag-helpers.js';
