+++ src/utils/skills-view-renderer.js
@@ -4,7 +4,8 @@
  * Pure DOM-construction functions. State is passed in, never owned here.
  */
 
-import { _el, renderList } from './dom.js';
+import { _el } from './dom-core.js';
+import { renderList } from './dom-lists.js';
 import { buildViewHeader } from './view-header.js';
 
 /**
